rename header.tsx to follow naming convention
