+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import "@/styles/globals.css";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import Header from "./_components/header";
+import Header from "./_components/Header";
 import { Providers } from "./providers"; // Import Providers component
 
 const geistSans = Geist({
